% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbreviate_factor.R
\name{abbreviate_factor}
\alias{abbreviate_factor}
\title{Abbreviate a factor.}
\usage{
abbreviate_factor(x, n = 1e+09)
}
\arguments{
\item{x}{factor}

\item{n}{numeric positive integer passed to abbreviate().}
}
\value{
a factor
}
\description{
Calls the base::abbreviate() function on a factor, and returns
a new factor with abbreviated level values.
}
\examples{
   x <- factor( c( 'catttt', 'dogggg', 'catttt' ) )
   out <- maeve:::abbreviate_factor( x, n = 3 )

}
\seealso{
\code{\link{factor}}
}
\author{
Bill Forrest \email{forrest@gene.com}
}
\keyword{abbreviate}
\keyword{factor}
