% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximate_first_derivative.R
\name{approximate_first_derivative}
\alias{approximate_first_derivative}
\title{Approximate the first derivative of a spline passing through a set of (x,y) pairs.}
\usage{
approximate_first_derivative(
  x,
  y,
  relative_tolerance = 0.01,
  dx_frac = 0.01,
  extended_output = FALSE
)
}
\arguments{
\item{x}{numeric abcissa values for the spline.  Should be}

\item{y}{numeric ordinae values for the spline.}

\item{relative_tolerance}{numeric tolerance parameter for how close to uniformly spaced the x-values must be.}

\item{dx_frac}{numeric relative value for how far out to step from given x-values in approximating first derivative.}

\item{extended_output}{logical. Whether to return a data frame with all the components leading to the first derivative.}
}
\value{
A numeric vector the same length as x & y with the first derivative of a spline through the points, or a data.frame if extended_output == TRUE.
}
\description{
Approximate the first derivative of a spline passing through a set of (x,y) pairs.
}
\examples{
 x <- seq(0, 2*pi, length = 101); 
 y <- sin(x) 
 dy.dx <- maeve:::approximate_first_derivative(x,y); 
 results <- cbind( x, y, cos(x), dy.dx ) # 3rd & 4th columns should match closely

}
\references{
\url{www.r-project.org}
}
\author{
Bill Forrest <forrest@gene.com>

Bill Forrest \email{forrest@gene.com}
}
\keyword{list}
