% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_parallel.R
\name{bootstrap_parallel}
\alias{bootstrap_parallel}
\title{Employ *parallel* parametric bootstrap to estimate rapidly the variation in summary efficacy metrics.}
\usage{
bootstrap_parallel(
  model_list,
  clean_DF,
  reference_Dunnett = NULL,
  full_inverse_function = NULL,
  metric = c("AUC", "ITGR", "linear", "AUC_pwl", "ITGR_pwl", "AUC_poly", "ITGR_poly"),
  progress = FALSE,
  bootstrap_statistic = c("AUC_originalScale", "Effect_Duration"),
  bootstrap_behavior = c("fixed", "random"),
  bootstrap_covariance_scale = 1,
  design_matrix = NULL,
  integral_grid_n = 101,
  xmin = NULL,
  xmax = NULL,
  x_time_spacing = 0.25,
  x_min_count_distinct = 10,
  norm_to_xrange = TRUE,
  K_boot = 1
)
}
\arguments{
\item{model_list}{list from maeve::model_study().}

\item{clean_DF}{data.frame.  It should be "model_list$data$clean_DF_pred".}

\item{reference_Dunnett}{character string with exact match to an element of grpnames.  The exact match will be the reference group in Dunnett contrasts.}

\item{full_inverse_function}{R function that inverts the response analyzed back to its original scale.}

\item{metric}{character string describing which response curve summary metric to use.}

\item{progress}{logical whether to print progress messages.}

\item{bootstrap_statistic}{character string:  Statistic for which the parametric bootstrap will assess uncertainty.}

\item{bootstrap_behavior}{character string:  Use fixed or random coefficient draws for parametric bootstrap?  The "fixed" option is intended for testing.}

\item{bootstrap_covariance_scale}{numeric scalar by which to scale multivariate Gaussian covariance matrix.}

\item{design_matrix}{matrix of numerics with properly ordered design matrix from "predict( md, type = 'lpmatrix' )", where "md" is model extracted from model_list.  This is found automatically if left as NULL (the default), but can be supplied by the user to speed up repeated execution, since it is typically unchanging within any given bootstrap simulation.}

\item{integral_grid_n}{integer number of grid points for numerical integration.}

\item{xmin}{numeric lower bound of range over which to evaluate the spline.}

\item{xmax}{numeric upper bound of range over which to evaluate the spline.}

\item{x_time_spacing}{spacing of time points across which to evaluate the splines.  If time values are recorded as integers, then the default should be sufficient.}

\item{x_min_count_distinct}{numeric integer.  If the number of distinct x-values at which splines are evaluated is less than this, a warning is issued.}

\item{norm_to_xrange}{logical for AUC calcuations: whether to normalize numeric definite integral to x-range.}

\item{K_boot}{integer number of bootstrap samples to draw.}
}
\value{
An R data.frame with a number of summary metrics (one value per treatment group).
}
\description{
Employ *parallel* parametric bootstrap to estimate rapidly the variation in summary efficacy metrics.
}
\examples{
 cat('Currently no working example for bootstrap_parallel().')

}
\references{
\url{www.r-project.org}
}
\author{
Bill Forrest <forrest@gene.com>

Bill Forrest \email{forrest@gene.com}
}
\keyword{list}
