% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_option_names.R
\name{check_option_names}
\alias{check_option_names}
\title{Check optional functional argument parameter names.}
\usage{
check_option_names(...)
}
\arguments{
\item{...}{optional parameters passed to a function.}
}
\value{
a vector the same length as the input vector, with either logical or numeric entries.
}
\description{
Check a list of optional "..." functional argument parameter names to confirm that they are all present in maeve_options().  If any are not, it may indicate, e.g., typos.  Take action depending on the value of maeve_options('mismatch_action').
}
\examples{
   cat('no example for check_option_names().')

}
\seealso{
\code{\link{setdiff}}
}
\author{
Bill Forrest \email{forrest@gene.com}
}
\keyword{setdiff}
