% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_design_matrix_from_basis.R
\name{construct_design_matrix_from_basis}
\alias{construct_design_matrix_from_basis}
\title{Make a design matrix from a grouping factor and numeric time vector from a user-selected basis type.}
\usage{
construct_design_matrix_from_basis(
  study_data_frame,
  basis_choice = c("tent", "poly"),
  format_digits = 3,
  use_poly_coefs = FALSE,
  group_name = maeve_options("group_name"),
  x_name = maeve_options("x_name"),
  break_points = maeve_options("break_points"),
  number_break_points = maeve_options("number_break_points"),
  poly_degree = maeve_options("poly_degree")
)
}
\arguments{
\item{study_data_frame}{data.frame with time ('x') vector and grouping vector}

\item{basis_choice}{character describing which basis function type to use in making a design matrix.}

\item{format_digits}{how many digits to retain in column names?}

\item{use_poly_coefs}{logical whether to use "attr( maeve_options('poly_object'), 'coef')" to scale poly basis matrix.}

\item{group_name}{character.      See ?maeve_options().}

\item{x_name}{character.      See ?maeve_options().}

\item{break_points}{numeric vector. See ?maeve_options().}

\item{number_break_points}{numeric.        See ?maeve_options().}

\item{poly_degree}{numeric.        See ?maeve_options().}
}
\value{
a design matrix, including intercept(s), that can be used with lme4::lmer().
}
\description{
Takes a data frame with a grouping factor and a vector of time 'x'-values and a
user choice of basis function type (currently 'tent' or 'poly') and returns
a design matrix for fitting a piecewise linear mixed effects model.
Tent basis vectors are created from the so-called "tent-functions".  See, e.g., Simon Wood (2017)
GAM 2nd edition text, section 4.2, page 164-166.  Polynomial basis vectors (not to be
confused with the gam spline approach) use the stats::poly() function.  This is
intended to be used in fitting linear mixed effects models to three (maybe four?) time points.
It can in principal be used for more than that, but in that circumstance the spline approach
will typically be better and is recommended.
}
\examples{
   ### Make a basis for piecewise linear regression over 21 days from the vismodegib data:
      require( magrittr ) # want pipe operator "\%>\%"
      vismo_pwl_basis <-
        vismodegib \%>\%
        dplyr::filter( DAY_OF_STUDY <= 21 ) \%>\%
        maeve:::construct_design_matrix_from_basis( basis_choice = 'tent', 
                                                    break_points = c( 0, 7, 14, 21 )
                                                   )

}
\seealso{
\code{\link{matrix}}
}
\author{
Bill Forrest \email{forrest@gene.com}
}
\keyword{matrix}
