% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_gam_formula.R
\name{construct_gam_formula}
\alias{construct_gam_formula}
\title{Make a gam formula with the basis_vec number provided.}
\usage{
construct_gam_formula(
  basis_vecs_value = -1,
  before_basis_vecs_value = "y ~ group +  s( x, bs = \\"tp\\", k =",
  after_basis_vecs_value = ", by = group )"
)
}
\arguments{
\item{basis_vecs_value}{numeric integer value to be used for number basis_vecs.}

\item{before_basis_vecs_value}{character string with the gam formula before the basis_vec number.}

\item{after_basis_vecs_value}{character string with the gam formula  after the basis_vec number.}
}
\value{
a formula object that can be passed to gamm4::gamm4(), which calls mgcv::gam().
}
\description{
Takes a number for the spline basis_vec value and pastes it
together with character strings giving the parts of the
formula before and after the basis_vec number, then returns the
pasted string as a formula.  This utility function circumvents
some specific function scope issues that arise when fitting
the gam model with a passed parameter number of basis_vecs.
}
\examples{
  ## avoid a ':::'-related warning in package check by not running next lines:
  \dontrun{
  test_gam_formula <- maeve:::construct_gam_formula( 6 )
  }

}
\seealso{
\code{\link{character}}
}
\author{
Bill Forrest \email{forrest@gene.com}
}
\keyword{character}
