% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_Effect_Duration.R
\name{determine_Effect_Duration}
\alias{determine_Effect_Duration}
\title{Determine the duration of effect for each treatment}
\usage{
determine_Effect_Duration(
  XY,
  colname_deriv1 = "pred_gam_deriv1",
  colname_ref_deriv1 = "base_pred_gam_deriv1",
  diff_threshold = -1e-04,
  grid_n = 1000,
  auto_round = TRUE,
  round_to = 4
)
}
\arguments{
\item{XY}{data.frame with times and original-scale tumor burdens, etc., for each of several identifiers.  Typically this is from the cleaned output data of maeve::model_study().}

\item{colname_deriv1}{character name of column with first derivative from GAM model fit.}

\item{colname_ref_deriv1}{character name of column with first derivative from specified reference group of GAM model fit.}

\item{diff_threshold}{numeric threshold to differentiate from zero, i.e., values below this are treated as zeros in evaluating changes.}

\item{grid_n}{numeric number of equally spaced grid points across which to approximate the effect duration.}

\item{auto_round}{logical: automatically round results to "round_to" digits?}

\item{round_to}{numeric number of digits to which the effect duration should be rounded.}
}
\value{
A numeric sum total of Partial Response cases recorded in the group.
}
\description{
Determine the duration of effect for each treatment
}
\examples{
 cat('Currently no working example for unexported function determine_Effect_Duration().')

}
\references{
\url{www.r-project.org}
}
\author{
Bill Forrest <forrest@gene.com>

Bill Forrest \email{forrest@gene.com}
}
\keyword{data.frame}
