% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_PR.R
\name{determine_PR}
\alias{determine_PR}
\title{Determine whether each identifier has a complete response (CR) at the end of study (EOS).}
\usage{
determine_PR(YY, min_val, PR_threshold, truncate_at_last_fit_day = FALSE)
}
\arguments{
\item{YY}{data.frame with times and original-scale tumor burdens, etc., for each of several identifiers.  Typically this is from the cleaned output data of maeve::fit_longitudinal_models().}

\item{min_val}{numeric threshold below which a tumor size is classified as a Complete Response (CR).  CRs by this criterion get excluded from the PR count.}

\item{PR_threshold}{numeric threshold below which a tumor size is classified as a Partial Response (PR).}

\item{truncate_at_last_fit_day}{logical: Use days only for which the group had a model fit.}
}
\value{
A numeric sum total of Partial Response cases recorded in the group.
}
\description{
Determine whether each identifier has a complete response (CR) at the end of study (EOS).
}
\examples{
 cat('Currently no working example for unexported function determine_PR().')

}
\references{
\url{www.r-project.org}
}
\author{
Bill Forrest <forrest@gene.com>

Bill Forrest \email{forrest@gene.com}
}
\keyword{data.frame}
