% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_noise.R
\name{draw_noise}
\alias{draw_noise}
\title{Create a ggplot2 object with root-mean-square prediction error from the conditional residuals for each subject in a study.}
\usage{
draw_noise(
  rmse,
  facet_by = c("group_name", "fit"),
  exclude_NA = FALSE,
  ncol_value = maeve_options("ncol_value"),
  nrow_value = maeve_options("nrow_value"),
  geom_na_rm = maeve_options("geom_na_rm"),
  geom_point_size = maeve_options("geom_point_size"),
  alpha_value = maeve_options("alpha_value"),
  axis_text_x_size = maeve_options("axis_text_x_size"),
  axis_text_x_angle = maeve_options("axis_text_x_angle"),
  axis_text_x_hjust = maeve_options("axis_text_x_hjust"),
  axis_text_x_vjust = maeve_options("axis_text_x_vjust"),
  axis_text_y_size = maeve_options("axis_text_y_size"),
  strip_text_size = maeve_options("strip_text_size"),
  title_label = maeve_options("title_label")
)
}
\arguments{
\item{rmse}{data frame with factors for group and ID, and numerics for RMSE-by-ID from linear fit and spline fit.  These four columns *MUST* be names {"group", "ID", "RMSE_lin_resid_ID", "RMSE_gam_resid_ID" }, respectively.  Typically, this data.frame is the data.frame called 'RMSE_by_ID' returned from maeve::predict_study() by setting "return_list = TRUE" in the predict_study() call.}

\item{facet_by}{character vector to determine whether to facet by group_name (default) or fit method.}

\item{exclude_NA}{logical whether to drop missing RMSE values.  Most often, this is because one of the two methods ('linear' or 'spline') simply was not fit.}

\item{ncol_value}{numeric.  See ?maeve_options().}

\item{nrow_value}{numeric.  See ?maeve_options().}

\item{geom_na_rm}{logical.  See ?maeve_options().}

\item{geom_point_size}{numeric.  See ?maeve_options().}

\item{alpha_value}{numeric.  See ?maeve_options().}

\item{axis_text_x_size}{numeric.  See ?maeve_options().}

\item{axis_text_x_angle}{numeric.  See ?maeve_options().}

\item{axis_text_x_hjust}{numeric.  See ?maeve_options().}

\item{axis_text_x_vjust}{numeric.  See ?maeve_options().}

\item{axis_text_y_size}{numeric.  See ?maeve_options().}

\item{strip_text_size}{numeric.  See ?maeve_options().}

\item{title_label}{numeric.  See ?maeve_options().}
}
\value{
A ggplot2 object
}
\description{
Create a ggplot2 object with root-mean-square prediction error from the conditional residuals for each subject in a study.
}
\examples{
 cat('Example for draw_noise() in dontrun block.')
 \dontrun{
 data( vismodegib )
 vismo21 <- dplyr::filter( vismodegib, DAY_OF_STUDY <= 21 )
 pred_list = vismo21 \%>\% model_study() \%>\% predict_study( return_list = TRUE )
 rmse = pred_list[[ 'RMSE_by_ID' ]]
 figure_of_noise <- draw_noise( rmse )
}

}
\references{
\url{www.r-project.org}
}
\author{
Bill Forrest <forrest@gene.com>

Bill Forrest \email{forrest@gene.com}
}
\keyword{ggplot2}
