% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_overlay.R
\name{draw_overlay}
\alias{draw_overlay}
\title{Create a ggplot2 object with model overlay fits from a maeve study.}
\usage{
draw_overlay(
  data_frame,
  fit = c("spline", "linear", "piecewise", "poly", "none"),
  group_name = maeve_options("group_name"),
  x_name = maeve_options("x_name"),
  endpoint_name = maeve_options("endpoint_name"),
  linear_predictor = maeve_options("linear_predictor"),
  spline_predictor = maeve_options("spline_predictor"),
  piecewise_predictor = maeve_options("piecewise_predictor"),
  poly_predictor = maeve_options("poly_predictor"),
  axis_text_x_size = maeve_options("axis_text_x_size"),
  axis_text_x_angle = maeve_options("axis_text_x_angle"),
  axis_text_y_size = maeve_options("axis_text_y_size"),
  strip_text_size = maeve_options("strip_text_size"),
  legend_text_size = maeve_options("legend_text_size"),
  legend_position_char = maeve_options("legend_position_char"),
  title_label = maeve_options("title_label"),
  x_label = maeve_options("x_label"),
  y_label = maeve_options("y_label"),
  geom_na_rm = maeve_options("geom_na_rm")
)
}
\arguments{
\item{data_frame}{data frame with a factor for group, numerics for time on study and response, and fitted numeric response curves.}

\item{fit}{character vector with names of fitted values to include.}

\item{group_name}{character.   See ?maeve_options().}

\item{x_name}{character.   See ?maeve_options().}

\item{endpoint_name}{character.   See ?maeve_options().}

\item{linear_predictor}{character.   See ?maeve_options().}

\item{spline_predictor}{character.   See ?maeve_options().}

\item{piecewise_predictor}{character.   See ?maeve_options().}

\item{poly_predictor}{character.   See ?maeve_options().}

\item{axis_text_x_size}{numeric.     See ?maeve_options().}

\item{axis_text_x_angle}{numeric.     See ?maeve_options().}

\item{axis_text_y_size}{numeric.     See ?maeve_options().}

\item{strip_text_size}{numeric.     See ?maeve_options().}

\item{legend_text_size}{numeric.     See ?maeve_options().}

\item{legend_position_char}{character.   See ?maeve_options().}

\item{title_label}{character.   See ?maeve_options().}

\item{x_label}{character.   See ?maeve_options().}

\item{y_label}{character.   See ?maeve_options().}

\item{geom_na_rm}{logical.     See ?maeve_options().}
}
\value{
A ggplot2 object
}
\description{
Create a ggplot2 object with model overlay fits from a maeve study.
}
\examples{
 cat("Currently no running example for draw_overlay(). See next block.")
 \dontrun{
 data( vismodegib )
 vismo21 <- dplyr::filter( vismodegib, DAY_OF_STUDY <= 21 )
 pred_data_frame = vismo21 \%>\% model_study() \%>\% predict_study()
 figure_overlay <- draw_overlay( pred_data_frame )
}

}
\references{
\url{www.r-project.org}
}
\author{
Bill Forrest <forrest@gene.com>

Bill Forrest \email{forrest@gene.com}
}
\keyword{data.frame}
