% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_study.R
\name{draw_study}
\alias{draw_study}
\title{Create a ggplot2 object with raw longitudinal data and optional model fits.}
\usage{
draw_study(
  data_frame,
  fit = c("none", "linear", "spline", "piecewise", "poly")[1],
  facet = TRUE,
  group_name = maeve_options("group_name"),
  subject_ID = maeve_options("subject_ID"),
  x_name = maeve_options("x_name"),
  endpoint_name = maeve_options("endpoint_name"),
  linear_predictor = maeve_options("linear_predictor"),
  spline_predictor = maeve_options("spline_predictor"),
  piecewise_predictor = maeve_options("piecewise_predictor"),
  poly_predictor = maeve_options("poly_predictor"),
  linear_color = maeve_options("linear_color"),
  spline_color = maeve_options("spline_color"),
  piecewise_color = maeve_options("piecewise_color"),
  poly_color = maeve_options("poly_color"),
  linear_lwd = maeve_options("linear_lwd"),
  spline_lwd = maeve_options("spline_lwd"),
  piecewise_lwd = maeve_options("piecewise_lwd"),
  poly_lwd = maeve_options("poly_lwd"),
  nrow_value = maeve_options("nrow_value"),
  ncol_value = maeve_options("ncol_value"),
  axis_text_x_size = maeve_options("axis_text_x_size"),
  axis_text_x_angle = maeve_options("axis_text_x_angle"),
  axis_text_x_hjust = maeve_options("axis_text_x_hjust"),
  axis_text_x_vjust = maeve_options("axis_text_x_vjust"),
  axis_text_y_size = maeve_options("axis_text_y_size"),
  strip_text_size = maeve_options("strip_text_size"),
  legend_text_size = maeve_options("legend_text_size"),
  legend_position_char = maeve_options("legend_position_char"),
  title_label = maeve_options("title_label"),
  x_label = maeve_options("x_label"),
  y_label = maeve_options("y_label"),
  geom_na_rm = maeve_options("geom_na_rm"),
  geom_point_size = maeve_options("geom_point_size"),
  alpha_value = maeve_options("alpha_value")
)
}
\arguments{
\item{data_frame}{data frame with factors for group and ID, numerics for time on study and response.}

\item{fit}{character vector with names of fitted values to include.}

\item{facet}{logical.  Whether to facet plots by group.  Defaults to TRUE.}

\item{group_name}{character.   See ?maeve_options().}

\item{subject_ID}{character.   See ?maeve_options().}

\item{x_name}{character.   See ?maeve_options().}

\item{endpoint_name}{character.   See ?maeve_options().}

\item{linear_predictor}{character.   See ?maeve_options().}

\item{spline_predictor}{character.   See ?maeve_options().}

\item{piecewise_predictor}{character.   See ?maeve_options().}

\item{poly_predictor}{character.   See ?maeve_options().}

\item{linear_color}{character.   See ?maeve_options().}

\item{spline_color}{character.   See ?maeve_options().}

\item{piecewise_color}{character.   See ?maeve_options().}

\item{poly_color}{character.   See ?maeve_options().}

\item{linear_lwd}{character.   See ?maeve_options().}

\item{spline_lwd}{character.   See ?maeve_options().}

\item{piecewise_lwd}{character.   See ?maeve_options().}

\item{poly_lwd}{character.   See ?maeve_options().}

\item{nrow_value}{numeric.     See ?maeve_options().}

\item{ncol_value}{numeric.     See ?maeve_options().}

\item{axis_text_x_size}{numeric.     See ?maeve_options().}

\item{axis_text_x_angle}{numeric.     See ?maeve_options().}

\item{axis_text_x_hjust}{numeric.     See ?maeve_options().}

\item{axis_text_x_vjust}{numeric.     See ?maeve_options().}

\item{axis_text_y_size}{numeric.     See ?maeve_options().}

\item{strip_text_size}{numeric.     See ?maeve_options().}

\item{legend_text_size}{numeric.     See ?maeve_options().}

\item{legend_position_char}{character.   See ?maeve_options().}

\item{title_label}{character.   See ?maeve_options().}

\item{x_label}{character.   See ?maeve_options().}

\item{y_label}{character.   See ?maeve_options().}

\item{geom_na_rm}{logical.     See ?maeve_options().}

\item{geom_point_size}{numeric.     See ?maeve_options().}

\item{alpha_value}{numeric.     See ?maeve_options().}
}
\value{
A ggplot2 object
}
\description{
Create a ggplot2 object with raw longitudinal data and optional model fits.
}
\examples{
 cat('Example for draw_study() in dontrun code block.')
 \dontrun{
 data( vismodegib )
 vismo21 <- dplyr::filter( vismodegib, DAY_OF_STUDY <= 21 )
 pred_data_frame = vismo21 \%>\% model_study() \%>\% predict_study()
 figure_by_group <- draw_study( pred_data_frame, endpoint_name = 'y',
                                fit = 'spline', ncol_value = 5 )
}

}
\references{
\url{www.r-project.org}
}
\author{
Bill Forrest <forrest@gene.com>

Bill Forrest \email{forrest@gene.com}
}
\keyword{data.frame}
