% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_log_odds.R
\name{estimate_log_odds}
\alias{estimate_log_odds}
\title{Estimate the log odds ratio and stanard error for counts of treatment relative to reference}
\usage{
estimate_log_odds(
  count_data_frame,
  group_name_char = "group",
  reference,
  number_on_first_day_char = "N_first_day",
  count_desc_char = c("PR", "EOS_CR", "PR_or_EOS_CR")[1],
  additive_offset = 0.5,
  return_original = TRUE
)
}
\arguments{
\item{count_data_frame}{data.frame with one row per group, total number per group, and number of positive counts out of the total}

\item{group_name_char}{character name of column with group factor identifier.}

\item{reference}{character name of level to use as a reference; must be in levels of factor designated by group_name_char.}

\item{number_on_first_day_char}{character name of column with total number in a each group.}

\item{count_desc_char}{character name of column with number of positive cases in each group.}

\item{additive_offset}{numeric value to add to each count in a 2 x 2 table before computing log odds ratio.  Default gives the Haldane-Anscombe-Gart correction of adding 0.5 to every count.}

\item{return_original}{logical whether to return the original counts in the table with the estimated log_OR and SE_log_OR.}
}
\value{
A data.frame with log odds ratio (log_OR) and its standard error (SE_log_OR) for each non-reference group relative to reference.
}
\description{
Estimate the log odds ratio and stanard error for counts of treatment relative to reference
}
\examples{
 test_log_odds <- data.frame( group = factor( c( 'group_01', 'group_02', 'group_03' ) ),
                              N_first_day = rep( 10, 3 ),
                              PR = c( 0, 2, 6 )
                             )
 log_odds_est <- maeve:::estimate_log_odds( test_log_odds, reference = 'group_01' )

}
\references{
\url{www.r-project.org}
}
\author{
Bill Forrest <forrest@gene.com>

Bill Forrest \email{forrest@gene.com}
}
\keyword{data.frame}
