% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freeze_factor_levels.R
\name{freeze_factor_levels}
\alias{freeze_factor_levels}
\title{re-order factor levels in a data frame to their current observed order.}
\usage{
freeze_factor_levels(x, no_freeze = NULL)
}
\arguments{
\item{x}{a data frame.}

\item{no_freeze}{character vector.  Columns of "x" named in this vector will be returned unchanged.}
}
\value{
data frame of the same size, names, and types as "x".  Factor level orders may have changed.
}
\description{
Takes a data frame and looks at each column.  If the column is _not_ a factor,
it will be returned unchanged.  If the column is a factor, it will be replaced
with a factor having the same levels, but re-ordered to whatever their observed
current level in the data frame happens to be.  It re-sets all the factors except
those named in the "no_freeze" argument. Intended as an improvement on the older
"freezeFactorLevels()".
}
\examples{
 myDat <- data.frame(
             animal = factor( c('cat', 'dog', 'cat', 'bird', 'fish', 'dog', 'cat'),
                              levels = c('bird','cat','dog','fish')
                            ),
             covering = factor( c('hair','hair', 'hair', 'feathers', 'scales', 'hair', 'hair'),
                                levels = c('feathers','hair','scales')
             ),
             feet = c(4,4,4,2,0,4,4)
          ) # end of data.frame 'myDat'


### Freeze all factors
myNewDat <- freeze_factor_levels( myDat )
print( levels( myDat$animal ) )
print( levels( myNewDat$animal) )
print( levels( myDat$covering ) )
print( levels( myNewDat$covering) )


### Freeze one but not another
myNewDat2 <- freeze_factor_levels( myDat, no_freeze = 'covering' )
print( levels( myDat$animal ) )
print( levels( myNewDat2$animal) )
print( levels( myDat$covering ) )
print( levels( myNewDat2$covering) )


### Freeze neither factor
myNewDat3 <- freeze_factor_levels( myDat, no_freeze = c('animal','covering') )
print( levels( myDat$animal ) )
print( levels( myNewDat3$animal) )
print( levels( myDat$covering ) )
print( levels( myNewDat3$covering) )

}
\seealso{
\code{\link{factor}}
}
\author{
Bill Forrest \email{forrest@gene.com}
}
\keyword{factor}
