% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_Effect_Duration_summary_statistics.R
\name{generate_Effect_Duration_summary_statistics}
\alias{generate_Effect_Duration_summary_statistics}
\title{Create a list with Effect_Duration summary statistics}
\usage{
generate_Effect_Duration_summary_statistics(
  model_list,
  clean_DF,
  study_ID_value = "",
  metric = c("AUC", "ITGR", "linear"),
  progress = FALSE,
  xmin = NULL,
  xmax = NULL,
  norm_to_xrange = TRUE,
  extended_output = FALSE,
  CI_lower_quantile = 0.025,
  CI_upper_quantile = 0.975,
  x_time_spacing = 0.25,
  x_min_count_distinct = 10,
  K_boot = 1000
)
}
\arguments{
\item{model_list}{R list with model objects from maeve::model_study().}

\item{clean_DF}{data.frame with fitted data, usually from model_list$data$clean_DF_pred.}

\item{study_ID_value}{character string with five digit DIVOS study ID, OR a DivoStudy object.}

\item{metric}{character vector specifying what metric(s) to use in longitudinal modeling.}

\item{progress}{logical determining whether to send interim progress messages.}

\item{xmin}{numeric lower bound of definite integral}

\item{xmax}{numeric upper bound of definite integral}

\item{norm_to_xrange}{logical whether or not to normalize estimated effects to interval length.}

\item{extended_output}{logical whether to return a full range of inferred interim table values.}

\item{CI_lower_quantile}{numeric value in the range (0,1) denoting the lower quanitle for confidence intervals.}

\item{CI_upper_quantile}{numeric value in the range (0,1) denoting the upper quanitle for confidence intervals.}

\item{x_time_spacing}{spacing of time points across which to evaluate the splines.  If time values are recorded as integers, then the default should be sufficient.}

\item{x_min_count_distinct}{numeric integer.  If the number of distinct x-values at which splines are evaluated is less than this, a warning is issued.}

\item{K_boot}{integer number of bootstrap samples to draw.}
}
\value{
An R list with named output.
}
\description{
Create a list with Effect_Duration summary statistics
}
\examples{
 cat('No example for unexported function generate_Effect_Duration_summary_statistics().')

}
\references{
\url{www.r-project.org}
}
\author{
Bill Forrest <forrest@gene.com>

Bill Forrest \email{forrest@gene.com}
}
\keyword{data.frame}
