% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_summary_table.R
\name{generate_summary_table}
\alias{generate_summary_table}
\title{Create a summary table making Dunnett comparisons.}
\usage{
generate_summary_table(
  model_list,
  clean_DF,
  reference_Dunnett = maeve_options("reference_Dunnett"),
  full_inverse_function = switch(maeve_options("inv_func_char"), Identity = function(x)
    {     x - maeve_options("add_to_endpoint") }, function(x) {    
    (get(maeve_options("inv_func_char")))(x) - maeve_options("add_to_endpoint") }),
  study_ID_value = "",
  metric = c("AUC", "ITGR", "linear", "AUC_pwl", "ITGR_pwl", "AUC_poly", "ITGR_poly"),
  progress = maeve_options("progress"),
  xmin = NULL,
  xmax = NULL,
  xrange_norm_method = maeve_options("xrange_norm_method"),
  extended_output = FALSE,
  conf_int = c("none", "TGI", "Effect_Duration")[1],
  EOS_CR_minval = maeve_options("EOS_CR_minval"),
  PR_threshold = maeve_options("PR_threshold"),
  CI_lower_quantile = 0.025,
  CI_upper_quantile = 0.975,
  x_time_spacing = 0.25,
  x_min_count_distinct = 10,
  K_boot = 1000
)
}
\arguments{
\item{model_list}{R list with model objects from maeve::model_study().}

\item{clean_DF}{data.frame with fitted data from maeve::model_study(), e.g., "model_list$data$clean_DF_pred".}

\item{reference_Dunnett}{character string with exact match to an element of grpnames.  The exact match will be the reference group in Dunnett contrasts.}

\item{full_inverse_function}{R function that inverts the response analyzed back to its original scale.}

\item{study_ID_value}{character string with five digit DIVOS study ID, OR a DivoStudy object.}

\item{metric}{character vector specifying what metric(s) to use in longitudinal modeling.}

\item{progress}{logical determining whether to send interim progress messages.}

\item{xmin}{numeric lower bound of definite integral}

\item{xmax}{numeric upper bound of definite integral}

\item{xrange_norm_method}{character string determining *how* the xrange normalization within leia() should be done.  Must be one of "none", "xrange", "slope_equivalent".}

\item{extended_output}{logical whether to return a full range of inferred interim table values.}

\item{conf_int}{character vector designating for which quantities a parametric bootstrap should be run to get confidence intervals.  Allowed options are "none" (the default", "TGI", "Effect_Duration", or c("TGI", "Effect_Duration").}

\item{EOS_CR_minval}{numeric passed to internal function to decide what constitutes a "complete response" for the the End-Of-Study Complete Response (EOS_CR) value on the original, untransformed response scale.}

\item{PR_threshold}{numeric passed to internal function to decide what constitutes a "partial response" for the PR value.  Fractional reduction is computed by-ID, from an ID-specific baseline value on the original, untransformed scale.}

\item{CI_lower_quantile}{numeric value in the range (0,1) denoting the lower quanitle for confidence intervals.}

\item{CI_upper_quantile}{numeric value in the range (0,1) denoting the upper quanitle for confidence intervals.}

\item{x_time_spacing}{spacing of time points across which to evaluate the splines.  If time values are recorded as integers, then the default should be sufficient.}

\item{x_min_count_distinct}{numeric integer.  If the number of distinct x-values at which splines are evaluated is less than this, a warning is issued.}

\item{K_boot}{integer number of bootstrap samples to draw.}
}
\value{
An R list with named output.
}
\description{
This is included principally for its options to generate legacy statistics, and is
not recommended for general use.
}
\examples{
 cat('Currently no working example for generate_summary_table().')

}
\references{
\url{www.r-project.org}
}
\author{
Bill Forrest <forrest@gene.com>

Bill Forrest \email{forrest@gene.com}
}
\keyword{data.frame}
