% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lmer_linfct.R
\name{get_lmer_linfct}
\alias{get_lmer_linfct}
\title{Return a matrix of linear contrasts, intended for use in multcomp::glht().}
\usage{
get_lmer_linfct(
  lmer_obj,
  contrast = c("Identity", "Dunnett", "Tukey", "Sequen", "custom"),
  grpnames = NULL,
  reference_Dunnett = NULL,
  slope_ind_char = ":x",
  custom_contrast = NULL
)
}
\arguments{
\item{lmer_obj}{an lmerMod object returned from lme4::lmer().}

\item{contrast}{character string specifying which type of contrast is desired.}

\item{grpnames}{character vector with names of levels for group factor.}

\item{reference_Dunnett}{character string with exact match to an element of grpnames.  The exact match will be the reference group in Dunnett contrasts.}

\item{slope_ind_char}{character string indicating how to discern slope parameters from "names(fixef(lmer_obj))".}

\item{custom_contrast}{numeric matrix of contrasts for slope estimates. Column names must match exact with treatment group names.  If it is not NULL, it overrides whatever is specified in "contrast".}
}
\value{
a contrast matrix.
}
\description{
Takes an lmerMod object from lme4::lmer() assuming a particular model formula
type, then finds a contrast matrix to query the slope estimates. The most
written is 'y ~  group + group:x + ( 1 + x | ID )'.
}
\examples{
 cat('No working example for unexported function get_lmer_linfct().')

}
\seealso{
\code{\link{data.frame}}
}
\author{
Bill Forrest \email{forrest@gene.com}
}
\keyword{lme4}
