% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xrange.R
\name{get_xrange}
\alias{get_xrange}
\title{Return x-range for a given group}
\usage{
get_xrange(
  dat,
  x_name = maeve_options("x_name"),
  group_name = maeve_options("group_name"),
  reference_Dunnett = maeve_options("reference_Dunnett")
)
}
\arguments{
\item{dat}{a data.frame including a numeric column and a factor column}

\item{x_name}{character string with name of numeric column of interest}

\item{group_name}{character string with name of factor column of interest}

\item{reference_Dunnett}{character string with exact match to an element of grpnames.  The exact match will be the reference group in Dunnett contrasts.}
}
\value{
a vector with two numbers giving the range of interest for dat[,x_name].
}
\description{
Takes a data frame with groups and a continuous abscissa, then returns the abscissa range for a given group.
}
\examples{
 test_dat <- expand.grid( x = 0:3, group = c('group_1','group_2') )
 example_xrange <- maeve:::get_xrange( test_dat, 'x', 'group' )

}
\seealso{
\code{\link{data.frame}}
}
\author{
Bill Forrest \email{forrest@gene.com}
}
\keyword{character}
