% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_NA_covariate.R
\name{impute_NA_covariate}
\alias{impute_NA_covariate}
\title{Impute / copy over missing covariate values in a data.frame}
\usage{
impute_NA_covariate(XX, colname_01, colname_02, stop_if_both_missing = TRUE)
}
\arguments{
\item{XX}{data.frame with two columns to be copied one from another}

\item{colname_01}{character name of column with possibly NA values.}

\item{colname_02}{character name of column with possibly NA values.}

\item{stop_if_both_missing}{logical whether to stop() if both entries are missing.}
}
\value{
An R data.frame with identical columns under colname_01 & colname_02.
}
\description{
Impute / copy over missing covariate values in a data.frame
}
\examples{
 test_levels <- c( 'one', 'two', 'three', 'four' )
 DF_with_NA <-
  data.frame( First  = factor( c(    NA, 'two', NA, 'four' ), levels = test_levels ),
              Second = factor( c( NA,  NA, 'three', NA ), levels = test_levels ),
              Third = factor( test_levels, levels = test_levels )
             )
 DF_imputed_12 <- maeve:::impute_NA_covariate( DF_with_NA, 'First', 'Second',
                                               stop_if_both_missing = FALSE )
 DF_imputed_13 <- maeve:::impute_NA_covariate( DF_with_NA, 'First', 'Third'  )

}
\references{
\url{www.r-project.org}
}
\author{
Bill Forrest <forrest@gene.com>

Bill Forrest \email{forrest@gene.com}
}
\keyword{data.frame}
