% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_time_to_threshold.R
\name{interpolate_time_to_threshold}
\alias{interpolate_time_to_threshold}
\title{Interpolate the first time at which a group's fitted curve increases by a given threshold over its baseline.}
\usage{
interpolate_time_to_threshold(
  YY,
  metric,
  full_inverse_function,
  threshold = 2,
  grid_n = 1000,
  auto_round = TRUE,
  round_to = 1
)
}
\arguments{
\item{YY}{a one-group subset of the 'clean_DF_pred' data frame returned from maeve::fit_longitudinal_models().}

\item{metric}{character vector specifying what metric(s) to use in longitudinal modeling.}

\item{full_inverse_function}{R function inverting the data transformation, taking the fitted values from YY back to the original units.}

\item{threshold}{numeric fold-change (in the original units, after "full_inverse_function()") from baseline that is desired to detect.}

\item{grid_n}{numeric number of points in the grid for interpolation.}

\item{auto_round}{logical: whether to automatically round the interpolated time to precision specified by "round_to"}

\item{round_to}{numeric number of digits to which we should round the interpolated progression time.}
}
\value{
A numeric progression time.
}
\description{
Interpolate the first time at which a group's fitted curve increases by a given threshold over its baseline.
}
\examples{
 cat('No working example for unexported function interpolate_time_to_threshold().')

}
\references{
\url{www.r-project.org}
}
\author{
Bill Forrest <forrest@gene.com>

Bill Forrest \email{forrest@gene.com}
}
\keyword{data.frame}
