% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lead_char.R
\name{lead_char}
\alias{lead_char}
\title{Add a leading character to equalize string lengths}
\usage{
lead_char(x, padChar = "0", padExtra = 0)
}
\arguments{
\item{x}{Vector of character, numeric, or factor values to pad.  Will be converted immediately to a character vector.}

\item{padChar}{character to add to the front of values of 'x'.  Should be of length 1, or unexpected behavior may result.}

\item{padExtra}{nonnegative integer of how many extra padChar characters to add to the front of values of 'x'.}
}
\value{
character string
}
\description{
Add a leading character to equalize string lengths
}
\examples{
lead_char( 1:10 )
lead_char( 1:10, padChar = '_' )
data.frame( animal = lead_char( c('cat','alligator', 'heron'), padChar = '_' ) )

}
\seealso{
\code{\link{mapply}}
}
\author{
Bill Forrest \email{forrest@gene.com}
}
\keyword{character}
