% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{maeve_defaults}
\alias{maeve_defaults}
\title{Return default valus of global options for maeve}
\usage{
maeve_defaults(...)
}
\arguments{
\item{...}{either comma-spaced option names as character strings, a concatenated character vector of such character strings, or integer positions in the option list.}
}
\value{
A list with maeve package defaults values for named parameters, but without resetting any option values.  If no names are specified, then the entire list of defaults is returned.
}
\description{
The function accepts either comma-spaced option names as character strings,
a concatenated character vector of such names, or integer positions in the
option list. It returns the default values corresponding to the input, without
resetting the existing option values.
}
\examples{
 cat( paste( 'maeve::maeve_defaults() returns default parameter',
             'values without resetting the current options.'
            ) 
    )
 \dontrun{ maeve_defaults( 'spline_color' );
           maeve_defaults( 'spline_color', 'alpha_value' );
           maeve_defaults( c( 'spline_color', 'alpha_value' ) );
           maeve_defaults( 1:3 ); # return defaults for first 3 components of maeve_options().
           ##
           maeve_options( alpha_value = 0.5 ) # change something.
           maeve_defaults( 'alpha_value' ); #
           ## Observe that the newly set value persists until maeve_reset() is called.
           maeve_options( 'alpha_value' )
          }

}
\author{
Bill Forrest \email{forrest@gene.com}
}
