% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{maeve_reset}
\alias{maeve_reset}
\title{Reset global options for maeve}
\usage{
maeve_reset()
}
\value{
resets the MAEVE_OPTIONS() internal function to its package defaults.
}
\description{
Reset global options for maeve
}
\examples{
 cat('maeve::maeve_options() allows reset of parameters called in functions within maeve.')
 \dontrun{ maeve_options( 'spline_color' );
           maeve_options( 'spline_color' = 'orange');
           maeve_options( 'spline_color' );
           maeve_reset()
           maeve_options( 'spline_color' );
          }

}
\author{
Bill Forrest \email{forrest@gene.com}
}
