% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_confint.R
\name{make_confint}
\alias{make_confint}
\title{Estimate confidence intervals for contrasts from multcomp::glht()}
\usage{
make_confint(
  glht_obj_name,
  glht_obj_list,
  conf_level = 0.95,
  adjustment_method = c("single-step", "Shaffer", "Westfall", "free", "holm",
    "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"),
  extended_output = FALSE,
  round_estimate = TRUE
)
}
\arguments{
\item{glht_obj_name}{character string with name of list component containing the glht object.}

\item{glht_obj_list}{list containing a glht object.}

\item{conf_level}{numeric. What level of confidence intervals should be returned?  Passed directly to stats::confint().}

\item{adjustment_method}{character string specifying a method for p-value adjustment.  Setting to 'none' will return unadjusted p-values and confidence intervals.  Setting to anything else (the default is 'single_step') will implement a particular p-value adjustment from multcomp.  Although the p-value adjustment method varies for different options of this parameter, all the options other than 'none' will return the same FWER-adjusted confidence intervals implemented via setting "calpha = multcomp::adjusted_calpha()" in "multcomp::confint.glht()". Setting adjustment_method = 'none' will return confidence intervals with "calpha = multcomp::univariate_calpha()" in "multcomp:::confint.glht()".}

\item{extended_output}{logical determining whether to compute extra output to return to a DIVOS summary table.}

\item{round_estimate}{logical determining whether to pass two data.frames through maeve::round_numerics() before merging on a numeric column.}
}
\value{
a data.frame
}
\description{
Input is the name of the glht object and a list containing it. This
choice reflects how the function is usually used in the workflow.
}
\examples{
 cat('Currently no working examples for maeve:::make_confint().')

}
\seealso{
\code{\link{data.frame}}
}
\author{
Bill Forrest \email{forrest@gene.com}
}
\keyword{lme4}
