% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_study.R
\name{model_study}
\alias{model_study}
\title{Fit longitudinal models to and return a list with models, data frames, and miscellaneous.}
\usage{
model_study(
  study_data_frame,
  drop_truncated_levels = FALSE,
  return_all = FALSE,
  record_truncated_group_levels = TRUE,
  metric = maeve_options("metric"),
  weight_lmer_option = maeve_options("weight_lmer_option"),
  group_name = maeve_options("group_name"),
  subject_ID = maeve_options("subject_ID"),
  x_name = maeve_options("x_name"),
  endpoint_name = maeve_options("endpoint_name"),
  add_to_endpoint = maeve_options("add_to_endpoint"),
  trans_func_char = maeve_options("trans_func_char"),
  inv_func_char = maeve_options("inv_func_char"),
  test_func_x = maeve_options("test_func_x"),
  progress = maeve_options("progress"),
  abbreviate_n = maeve_options("abbreviate_n"),
  reference_Dunnett = maeve_options("reference_Dunnett"),
  restrict_x = maeve_options("restrict_x"),
  overall_x_min = maeve_options("overall_x_min"),
  overall_x_max = maeve_options("overall_x_max"),
  min_n_in_group = maeve_options("min_n_in_group"),
  min_frac_in_group = maeve_options("min_frac_in_group"),
  min_frac_in_study = maeve_options("min_frac_in_study"),
  number_basis_vecs = maeve_options("number_basis_vecs"),
  min_basis_vecs = maeve_options("min_basis_vecs"),
  max_basis_vecs = maeve_options("max_basis_vecs"),
  autoset_full_study_data = maeve_options("autoset_full_study_data"),
  autoset_modeling_data = maeve_options("autoset_modeling_data")
)
}
\arguments{
\item{study_data_frame}{a well-prepared data.frame with the study data and appropriate columns.}

\item{drop_truncated_levels}{logical: whether to drop from factors levels that have been truncated away in the modeling data.}

\item{return_all}{logical: return a whole bunch of output in a list.}

\item{record_truncated_group_levels}{logical: whether to record level(s) of any groups that are entirely lost during data truncation in maeve_options("truncated_group_levels").}

\item{metric}{character.      See ?maeve_options().}

\item{weight_lmer_option}{character.      See ?maeve_options().}

\item{group_name}{character.      See ?maeve_options().}

\item{subject_ID}{character.      See ?maeve_options().}

\item{x_name}{character.      See ?maeve_options().}

\item{endpoint_name}{character.      See ?maeve_options().}

\item{add_to_endpoint}{numeric.        See ?maeve_options().}

\item{trans_func_char}{character.      See ?maeve_options().}

\item{inv_func_char}{character.      See ?maeve_options().}

\item{test_func_x}{numeric.        See ?maeve_options().}

\item{progress}{logical.        See ?maeve_options().}

\item{abbreviate_n}{numeric.        See ?maeve_options().}

\item{reference_Dunnett}{character.      See ?maeve_options().}

\item{restrict_x}{logical.        See ?maeve_options().}

\item{overall_x_min}{numeric.        See ?maeve_options().}

\item{overall_x_max}{numeric.        See ?maeve_options().}

\item{min_n_in_group}{numeric.        See ?maeve_options().}

\item{min_frac_in_group}{numeric.        See ?maeve_options().}

\item{min_frac_in_study}{numeric.        See ?maeve_options().}

\item{number_basis_vecs}{numeric.        See ?maeve_options().}

\item{min_basis_vecs}{numeric.        See ?maeve_options().}

\item{max_basis_vecs}{numeric.        See ?maeve_options().}

\item{autoset_full_study_data}{logical.        See ?maeve_options().}

\item{autoset_modeling_data}{logical.        See ?maeve_options().}
}
\value{
An R list with named output.
}
\description{
Fit longitudinal models to and return a list with models, data frames, and miscellaneous.
}
\examples{
 cat('Example for model_study() in dontrun code block.')
 \dontrun{
 data( vismodegib )
 vismo21 <- dplyr::filter( vismodegib, DAY_OF_STUDY <= 21 )
 model_list <- model_study( vismo21 )
}

}
\references{
\url{www.r-project.org}
}
\author{
Bill Forrest <forrest@gene.com>

Bill Forrest \email{forrest@gene.com}
}
\keyword{data.frame}
