% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_in_list.R
\name{models_in_list}
\alias{models_in_list}
\title{Estimate linear mixed and / or gamm4 and/or piecewise linear and/or polynomial models and return in a list.}
\usage{
models_in_list(
  clean_DF,
  metric = c("linear", "AUC", "ITGR"),
  study_ID_value,
  progress,
  number_basis_vecs,
  min_basis_vecs,
  max_basis_vecs,
  weight_lmer_option = c("uniform", "overweight_baseline"),
  lme4_formula = formula("y ~ group + group:x + ( 1 + x | ID )"),
  gam_formula_char_list = list(prefix = "y ~ group +  s( x, bs = \\"tp\\", k =", suffix
    = ", by = group )"),
  allow_basis_vec_n_change = TRUE,
  allowed_number_fix_attempts = 20
)
}
\arguments{
\item{clean_DF}{data.frame with study data loaded and organized for the modeling call.}

\item{metric}{character vector specifying what metric(s) to use in longitudinal modeling.}

\item{study_ID_value}{character string with five digit DIVOS study ID, OR a DivoStudy object.}

\item{progress}{logical determining whether to send interim progress messages.}

\item{number_basis_vecs}{numeric: number of basis_vecs to include in spline basis used by mgcv}

\item{min_basis_vecs}{numeric lower bound on number of basis_vecs in spline basis.}

\item{max_basis_vecs}{numeric upper bound on number of basis_vecs in spline basis.}

\item{weight_lmer_option}{character ("uniform" or "overweight_baseline") specifying which of two weighting option schemes to use for linear modeling intercept.}

\item{lme4_formula}{formula for lmer() call.}

\item{gam_formula_char_list}{two component list of character strings providing the mgcv::gam() formula prefix & suffix to the basis_vec value.}

\item{allow_basis_vec_n_change}{logical allow the function to change the number of basis_vecs to avoid a crash?}

\item{allowed_number_fix_attempts}{numeric nonnegative integer.}
}
\value{
An R list with named output.
}
\description{
Estimate linear mixed and / or gamm4 and/or piecewise linear and/or polynomial models and return in a list.
}
\examples{
 cat('Currently no working example for models_in_list().')

}
\references{
\url{www.r-project.org}
}
\author{
Bill Forrest <forrest@gene.com>

Bill Forrest \email{forrest@gene.com}
}
\keyword{data.frame}
