% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_to_first_timepoint.R
\name{normalize_to_first_timepoint}
\alias{normalize_to_first_timepoint}
\title{Normalize a time-indexed vector of values its first value.}
\usage{
normalize_to_first_timepoint(
  DATA_FRAME,
  x_name = "x",
  y_name = "y",
  average_multiple_obs = TRUE,
  centering_value = 1,
  scale_value = 100
)
}
\arguments{
\item{DATA_FRAME}{data.frame with x-values (i.e, time-values) and y-values (responses.}

\item{x_name}{character string with name of time value that order the responses.}

\item{y_name}{character string with name of responses to be normalized to their own temporal baseline value.}

\item{average_multiple_obs}{logical should we allow averaging of multiple baseline values, or call an error?}

\item{centering_value}{numeric value to subtract from the normalized vector.}

\item{scale_value}{numeric value by which to re-scale the normalized, centered vector.}
}
\value{
A numeric vector with normalized values.
}
\description{
From a data.frame, extract a y-vector of responses and their x-vector of 
observation times.  Normalize the y-vector to its first (based on the
x-values) observation, subtract out the centering value from the entire
normalized vector, and return the normalized, centered vector.
}
\examples{
normalizing_test_data <- data.frame( x = seq(0,1,length = 11) )
normalizing_test_data$y <- 1 + 2 * exp(-normalizing_test_data$x)
normalized_output <-
  data.frame( normalizing_test_data,
              y_norm = maeve:::normalize_to_first_timepoint( normalizing_test_data )
             )

}
\references{
\url{www.r-project.org}
}
\author{
Bill Forrest <forrest@gene.com>

Bill Forrest \email{forrest@gene.com}
}
\keyword{data.frame}
