% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_study.R
\name{permute_study}
\alias{permute_study}
\title{Create a data frame with group_name assignments permuted randomly with respect to the subject_ID values.}
\usage{
permute_study(
  data_frame,
  random_seed = NULL,
  group_name = maeve_options("group_name"),
  subject_ID = maeve_options("subject_ID"),
  x_name = maeve_options("x_name")
)
}
\arguments{
\item{data_frame}{data frame with factors for group and ID, numerics for time on study and response.}

\item{random_seed}{numeric passed to base::set.seed()}

\item{group_name}{character.   See ?maeve_options().}

\item{subject_ID}{character.   See ?maeve_options().}

\item{x_name}{character.   See ?maeve_options().}
}
\value{
A data.frame
}
\description{
Create a data frame with group_name assignments permuted randomly with respect to the subject_ID values.
}
\examples{
 cat('Example for permute_study() in dontrun{} code block.')
 \dontrun{
 data( vismodegib )
 vismo30 <- dplyr::filter( vismodegib, DAY_OF_STUDY <= 30 )
 vismo30 <- dplyr::mutate( vismo30, y = log( 1 + TUMOR_VOLUME ) )
 vismo30_permuted <- permute_study( vismo30, random_seed = 12345 )
 ## see outcome of permutation:
 fig_vismo_orig <- draw_study( vismo30,          ncol_value = 5, endpoint_name = 'y' )
 fig_vismo_perm <- draw_study( vismo30_permuted, ncol_value = 5, endpoint_name = 'y' )
}

}
\references{
\url{www.r-project.org}
}
\author{
Bill Forrest <forrest@gene.com>

Bill Forrest \email{forrest@gene.com}
}
\keyword{data.frame}
