% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_study.R
\name{predict_study}
\alias{predict_study}
\title{Take fit longitudinal models and raw data and return data frames with various predicted values.}
\usage{
predict_study(
  model_list,
  clean_DF_restricted = NULL,
  clean_DF_full_xrange = NULL,
  return_list = FALSE,
  reference_Dunnett = maeve_options("reference_Dunnett"),
  modeling_data_frame = maeve_options("modeling_data_frame"),
  full_study_data_frame = maeve_options("full_study_data_frame"),
  metric = maeve_options("metric"),
  progress = maeve_options("progress"),
  contrast = maeve_options("contrast"),
  truncate_fit = maeve_options("truncate_fit"),
  group_name = maeve_options("group_name"),
  subject_ID = maeve_options("subject_ID"),
  x_name = maeve_options("x_name"),
  endpoint_name = maeve_options("endpoint_name"),
  x_pred_type = maeve_options("x_pred_type"),
  x_pred_vec = maeve_options("x_pred_vec"),
  x_pred_spacing = maeve_options("x_pred_spacing"),
  x_pred_interior_grid = maeve_options("x_pred_interior_grid"),
  include_newdata_ID = maeve_options("include_newdata_ID"),
  N_integration_grid = maeve_options("N_integration_grid"),
  add_to_endpoint = maeve_options("add_to_endpoint"),
  trans_func_char = maeve_options("trans_func_char"),
  inv_func_char = maeve_options("inv_func_char"),
  test_func_x = maeve_options("test_func_x")
)
}
\arguments{
\item{model_list}{list with two components returned from maeve:::models_in_list() and/or maeve::model_study().}

\item{clean_DF_restricted}{data frame with data restricted to those used in fitting the models.  Most often set to 'clean_DF_restricted' data.frame returned from maeve::model_study().}

\item{clean_DF_full_xrange}{data frame including data beyond that used for the modeling.  Most often set to NULL or to the 'clean_DF_full_xrange' data.frame returned from maeve::model_study().}

\item{return_list}{logical: return a whole bunch of output in a list.}

\item{reference_Dunnett}{character.      See ?maeve_options().}

\item{modeling_data_frame}{data.frame.     See ?maeve_options().}

\item{full_study_data_frame}{data.frame.     See ?maeve_options().}

\item{metric}{character.      See ?maeve_options().}

\item{progress}{logical.        See ?maeve_options().}

\item{contrast}{character.      See ?maeve_options().}

\item{truncate_fit}{logical.        See ?maeve_options().}

\item{group_name}{character.      See ?maeve_options().}

\item{subject_ID}{character.      See ?maeve_options().}

\item{x_name}{character.      See ?maeve_options().}

\item{endpoint_name}{character.      See ?maeve_options().}

\item{x_pred_type}{character.      See ?maeve_options().}

\item{x_pred_vec}{numeric vector. See ?maeve_options().}

\item{x_pred_spacing}{numeric.        See ?maeve_options().}

\item{x_pred_interior_grid}{logical.        See ?maeve_options().}

\item{include_newdata_ID}{logical.        See ?maeve_options().}

\item{N_integration_grid}{numeric.        See ?maeve_options() for documentation.}

\item{add_to_endpoint}{numeric.        See ?maeve_options().}

\item{trans_func_char}{character.      See ?maeve_options().}

\item{inv_func_char}{character.      See ?maeve_options().}

\item{test_func_x}{numeric.        See ?maeve_options().}
}
\value{
An R data.frame OR an R list with several different data.frames & other objectd, depending on "return_list" value.
}
\description{
Take fit longitudinal models and raw data and return data frames with various predicted values.
}
\examples{
 cat('Example for predict_study() in dontrun code block.')
 \dontrun{
 data( vismodegib )
 vismo21 <- dplyr::filter( vismodegib, DAY_OF_STUDY <= 21 )
 model_list <- model_study( vismo21 )
 pred_data_frame = predict_study( model_list )
}

}
\references{
\url{www.r-project.org}
}
\author{
Bill Forrest <forrest@gene.com>

Bill Forrest \email{forrest@gene.com}
}
\keyword{data.frame}
