% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_of_instant_dropout.R
\name{prob_of_instant_dropout}
\alias{prob_of_instant_dropout}
\title{Compute probability of dropout just after a time point based on current endpoint value.}
\usage{
prob_of_instant_dropout(
  x_vector,
  y_vector,
  dropout_hazard_map = data.frame(y = seq(0, 2000, by = 200), drop_haz = c(rep(0.025,
    3), seq(0.1, 0.3, length = 7), log(2))),
  MIN_endpoint = min(dropout_hazard_map[, 1]),
  MAX_endpoint = max(dropout_hazard_map[, 1])
)
}
\arguments{
\item{x_vector}{numeric vector of time points at which endpoint values are observed.}

\item{y_vector}{numeric vector of endpoint values.}

\item{dropout_hazard_map}{data.frame with the first column a set of endpoint values spanning the relevant range of endpoints, the second column dropout probabilities corresponding to the endpoint values.}

\item{MIN_endpoint}{numeric  lowest value of endpoint to be evaluated.}

\item{MAX_endpoint}{numeric highest value of endpoint to be evaluated.}
}
\value{
a numeric vector the same length as y_vector.
}
\description{
Takes an input value 'y' which is an endpoint at a given time.  Returns a probability
interpolated based on a spline interpolating paired { y_vector_i = endpoint_i, dropout_hazard_map_i = dropout_probability_i } pairs.
}
\examples{
cat('no current example for unexported function prob_of_instant_dropout.')
\dontrun{
prob <- prob_of_instant_dropout( c( 1:5, 10, 12, 20 ), c( rep(100,5), 2001, 1500, 800 ) )


}

}
\references{
\url{www.r-project.org}
}
\author{
Bill Forrest <forrest@gene.com>

Bill Forrest \email{forrest@gene.com}
}
\keyword{interpolation}
