% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpson_coefficients.R
\name{simpson_coefficients}
\alias{simpson_coefficients}
\title{Find coefficients used in Simpson's approximation for numerical integration}
\usage{
simpson_coefficients(
  x,
  return_list = FALSE,
  diff_TOL = 0.001,
  suppress_warning = TRUE
)
}
\arguments{
\item{x}{numeric vector of abcissa values in ascending order.}

\item{return_list}{logical whether to return computed coefficients in a vector or the separate components in a list.}

\item{diff_TOL}{numeric measure of how uneven the x-values can be (they should be evenly spaced).}

\item{suppress_warning}{logical for whether to return warning.}
}
\value{
vector of numeric cofficients if return_list is TRUE; list of integers coefficients and interval size "h" separately if return_list is FALSE
}
\description{
Takes an increasing, equally spaced numeric vector 'x' and returns 
Simpson's Rule cofficients used in numerical integration.
}
\examples{
 x <- seq(0, 1, by = .1); y <- 6 * x * ( 1 - x )
 w <- simpson_coefficients( x )
 print( sum( w * y ) )

 simpson_coefficients( x, return_list = TRUE )

}
\references{
\url{www.r-project.org}
}
\author{
Bill Forrest <forrest@gene.com>

Bill Forrest \email{forrest@gene.com}
}
\keyword{integration}
