% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splice_color_order.R
\name{splice_color_order}
\alias{splice_color_order}
\title{splice color ordering factor to rotate across groups.}
\usage{
splice_color_order(group, ID, as_factor = TRUE)
}
\arguments{
\item{group}{factor denoting group membership.}

\item{ID}{factor denoting identification. Levels must be nested within group (i.e., an ID cannot be listed in 2+ different groups).}

\item{as_factor}{logical return the result as a factor (the default) or, as an alternative, return the whole data frame.}
}
\value{
a factor
}
\description{
Takes equal-length factors for group membership and ID value (nested within group membership)
and returns a new factor ordering more evenly across groups.  E.g., for "K" groups,
the first level is the first ID in group #1, second level is the first ID in group #2,
..., the K'th level is the first ID in group #K, then the (K+1)st level is the second
ID in group #1, and so on.  The logic of the algorithm assumes that
}
\details{
identical( dat, dat[ order( group, ID ),] )

is TRUE, since this is the use case within which it was developed, 
and otherwise it's ambiguous what a user would expect. The function
will still run if this condition fails, but a warning is issued and 
results are unpredictable.
}
\examples{
 my_group <- factor( rep(paste('group',lead_char(1:3),sep = '_'),each = 4) )
 my_ID    <- factor( paste('ID',lead_char(1:12), sep = '_') )
 my_color_order <- splice_color_order ( my_group, my_ID )

}
\seealso{
\code{\link{factor}}
}
\author{
Bill Forrest \email{forrest@gene.com}
}
\keyword{factor}
