% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_columns.R
\name{swap_columns}
\alias{swap_columns}
\title{Swap positions (so names and values both) for two columns named in a data.frame.}
\usage{
swap_columns(XX, column_name_01, column_name_02)
}
\arguments{
\item{XX}{data.frame with two or more columns.}

\item{column_name_01}{character string with the name of first column.}

\item{column_name_02}{character string with the name of first column.}
}
\value{
a data.frame with the two named columns interchanged.
}
\description{
Takes a data frame and two character strings.  The character strings must
match exactly to the column names of two columns in the data frame. The
returned value is the same data frame, but with the column positions of 
the two columns named interchanged.
}
\examples{
  ### avoid a ':::'-related warning in package check by not running next lines:
   dat <- data.frame( A = letters[1:10], X = 1:10, Y = rnorm( 10 ) )
   dat_swap <- maeve:::swap_columns( dat, 'X', 'Y' )

}
\seealso{
\code{\link{data.frame}}
}
\author{
Bill Forrest \email{forrest@gene.com}
}
\keyword{data.frame}
