% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tally_study.R
\name{tally_study}
\alias{tally_study}
\title{Calculate group-level summaries based solely on raw data (i.e., that do not require any modeling).}
\usage{
tally_study(
  study_data_frame,
  group_name = maeve_options("group_name"),
  subject_ID = maeve_options("subject_ID"),
  x_name = maeve_options("x_name"),
  endpoint_name = maeve_options("endpoint_name"),
  reference = levels(study_data_frame[, group_name])[1],
  response = c("none", "PR", "EOS_CR", "PR_or_EOS_CR")[1],
  EOS_CR_minval = maeve_options("EOS_CR_minval"),
  PR_threshold = maeve_options("PR_threshold"),
  additive_offset = 0.5,
  na.rm = TRUE
)
}
\arguments{
\item{study_data_frame}{a well-prepared data.frame with the study data and appropriate columns.}

\item{group_name}{character name of treatment regimen factor}

\item{subject_ID}{character identifier for individual within a group.}

\item{x_name}{character name of independent variable (usually time or concentration).}

\item{endpoint_name}{character name of   dependent variable (e.g., tumor volume, body weight).}

\item{reference}{character name of designated reference group.  Must be a level from the factor named by group_name.}

\item{response}{character name of by-group count summary.}

\item{EOS_CR_minval}{numeric   passed to internal function to decide what constitutes a "complete response" for the the End-Of-Study Complete Response (EOS_CR) value on the original, untransformed response scale.}

\item{PR_threshold}{numeric   passed to internal function to decide what constitutes a "partial response" for the PR value.  Fractional reduction is computed by-ID, from an ID-specific baseline value on the original, untransformed scale.}

\item{additive_offset}{numeric   value to add to each count in a 2 x 2 table before computing log odds ratio.  Default gives the Haldane-Ascombe-Gart correction of adding 0.5 to every count.}

\item{na.rm}{logical   whether to remove NA values when computing summaries.}
}
\value{
An R data.frame with one row per group.
}
\description{
Calculate group-level summaries based solely on raw data (i.e., that do not require any modeling).
}
\examples{
 vismo_tally <- tally_study( vismodegib )

}
\references{
\url{www.r-project.org}
}
\author{
Bill Forrest <forrest@gene.com>

Bill Forrest \email{forrest@gene.com}
}
\keyword{data.frame}
