% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_final_colon.R
\name{trim_final_colon}
\alias{trim_final_colon}
\title{Remove any trailing ":" colon from each element of a character vector.}
\usage{
trim_final_colon(x)
}
\arguments{
\item{x}{character vector or NULL}
}
\value{
a character vector or NULL
}
\description{
Strips out a trailing colon (if it is present) from each element of a
character string.  This is to deal with the apparent (but troublesome)
fact that glht results from "lmerMod" do not add such a colon, but results
from the same function call on a "gam" object do, which was causing errors
when they were combined.
}
\examples{
 strings_with_colons <- c("cat:", "dog",  ":")
 
 trimmed_strings <- maeve:::trim_final_colon( strings_with_colons )

}
\seealso{
\code{\link{character}}
}
\author{
Bill Forrest \email{forrest@gene.com}
}
\keyword{character}
