% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncate_study.R
\name{truncate_study}
\alias{truncate_study}
\title{Truncate x-range for a data-frame based on x-axis values and optionally y-axis response values.}
\usage{
truncate_study(
  study_data_frame,
  group_name = maeve_options("group_name"),
  subject_ID = maeve_options("subject_ID"),
  x_name = maeve_options("x_name"),
  endpoint_name = NULL,
  fit = NULL,
  overall_x_min = maeve_options("overall_x_min"),
  overall_x_max = maeve_options("overall_x_max"),
  min_n_in_group = maeve_options("min_n_in_group"),
  min_frac_in_group = maeve_options("min_frac_in_group"),
  min_frac_in_study = maeve_options("min_frac_in_study"),
  print_warning = TRUE,
  droplevels = TRUE,
  drop_singletons = FALSE,
  drop_singleton_IDs = TRUE,
  singleton_number = 1,
  truncation_return_type = maeve_options("truncation_return_type"),
  reset_names = TRUE
)
}
\arguments{
\item{study_data_frame}{data.frame with data to filter.}

\item{group_name}{character with column name of treatment group factor.}

\item{subject_ID}{character with column name of subject ID factor.}

\item{x_name}{character with column name of x-axis numeric variable}

\item{endpoint_name}{optional character with column name of a y-axis numeric variable.}

\item{fit}{optional           character with column name of predicted values for a y-axis numeric variable.}

\item{overall_x_min}{numeric hard limit on the lower range of x-values.}

\item{overall_x_max}{numeric hard limit on the upper range of x-values.}

\item{min_n_in_group}{numeric x-values in a group will be truncated to last x-value when at least this number are still in the group.}

\item{min_frac_in_group}{numeric x-values in a group will be truncated to last x-value when at least this fraction are still in the group.}

\item{min_frac_in_study}{numeric x-values in a group will be truncated to last x-value when at least this fraction are still in the study..}

\item{print_warning}{logical whether to print warnings.}

\item{droplevels}{logical whether to restrict factors to levels present in the data frame provided.}

\item{drop_singletons}{logical whether to exclude *all* observations from a group if, after truncation, it has no more than one observation per subject.}

\item{drop_singleton_IDs}{logical whether to exclude *all* observations for one ID   if, after truncation, it has no more than one observation total.}

\item{singleton_number}{numeric If all IDs in a group have "singleton_number" or few observations and drop_singletons == TRUE, the group will get dropped.}

\item{truncation_return_type}{character string to describe output content.  Must be one of "logical", "data.frame", or "list".}

\item{reset_names}{logical whether to reset the column names to original values in optional output.}
}
\value{
a logical vector length equal to the number of rows in study_data_frame.
}
\description{
Takes a data frame and character identifiers for various columns; returns a logical vector the same length as the number of rows in the data frame.
}
\examples{
 cat('no working example yet')

}
\seealso{
\code{\link{data.frame}}
}
\author{
Bill Forrest \email{forrest@gene.com}
}
\keyword{truncate}
